Refactor authentication endpoints and implement user login and registration logic
- Updated router configuration to register login and registration controllers under the /auth prefix.
- Created loginController and registerController for handling user authentication, including input validation and error handling.
- Removed registration logic from userController to streamline responsibilities and enhance modularity.
