+++ src/controller/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as registerController } from './registerController';
+export { default as loginController } from './loginController';
