+++ src/router.ts
@@ -1,4 +1,5 @@
 import { FastifyInstance } from "fastify";
+import { loginController, registerController } from "./controller/auth";
 import userController from "./controller/userController";
 import indexController from "./controller/indexController";
 import priceController from "./controller/priceController";
@@ -7,7 +8,12 @@ import volumeController from "./controller/volumeController";
 
 export default async function router(fastify: FastifyInstance) {
   // Auth endpoints
-  fastify.register(userController, { prefix: "/auth" });
+  fastify.register(async (fastify) => {
+    fastify.register(loginController);
+    fastify.register(registerController);
+  }, { prefix: "/auth" });
+  
+  fastify.register(userController, { prefix: "/user" });
   
   // API endpoints
   fastify.register(indexController, { prefix: "/" });
